Port Accessibility for KateView to Qt5
Remove the cursor accessible as that is superceeded by the text
interface.

This brings the static text interface back, with accerciser the text is
visible, but changes are not yet reflected.

To test this is a bit tricky as a kde 5 session is required but that
means that the accessibility dbus is not available. I patched qtbase to
directly connect to the normal accessibility bus instead of trying to
use the session bus to find the other bus' address.