fix KTextEditor::TextHintInterface
This incompatible change is required to allow multiple text hint (tool
tip) providers at the same time. How to adapt to the new interface is
described in detail in texthintinterface.h.

Afaik, the only user of this interface is Sven in KDevelop.
CCMAIL: svenbrauch@googlemail.com
CCMAIL: aleixpol@kde.org