Refactor default-color generation into a central place.
This is a forward-port of commit c2dc123f624deff0098bdb74cf9db4542
in the KDE4 kate repository.

Instead of copy'n'pasting the generation of default colors, we put
that into a central place and generate the colors there.

Building on top of this patch, I plan to add some auto-tinting
code as I use in KDevelop for the currently hardcoded colors
(see Qt::yellow & friends).