normal mode and vi-mode separation
This changes hides vi-mode behind generic input mode interface, which
means that internals of ktexteditor no longer depends on vi-mode. There
are no "if"s and only single one #ifdef. And yes there is
a BUILDE_VIMODE cmake option now, which enables building ktexteditor
without the vi-mode.

==  Why ==

Reason for doing this is the reduction of coupling and removal of those
"if(viInputMode())" cases.

== How it does ==

There is a new directory "inputmode", where the interfaces are. The
KateAbstractInputModeFactory is ment to by initialized in KateGlobal.
The view can then use it and ask it for KateAbstractInputMode instance,
which it can then use in special places like "findNext" or
"handleKeyPress", etc. The KateViInputModeFactory now hides KateViGlobal
and KateViInputMode hides the KateViInputModeManager from the rest of
the system.

The interface of KateAbstractInputMode is currently in very bad shape,
but that can be cleaned up easily. Lots of the places are oneliners,
either changing because there is no KateView::getViModeManager anymore
or because there are now multiple input modes (and yes, it means that
adding EmacsInputMode after this would not be that much of a nightmare).

== Notes ==

As for the:
* relative line numbers - plan is to let IconBar call the InputMode for
  rendering that specific part
* vi-mode specific config options - they are left alone and can be
  refactored afterwards into KateViInputModeFactory
* input mode specific actions - I would like to see them in Factories
  only, but that might not play well with the shortcut config UI. For
  now I think it's not important.

REVIEW: 116770
GUI: changes menu "Vi Input Mode" actions