put right confusion from previous commit
just an empty commit to put master branch right (I hope)

this is what I think happenned:
my earlier commit with title ....

9436c17	2014-05-01	Trevor Blight	 add ada indenter to ktext editor

... just dissapeared (????).
After the immediately following commit (9df6631) it wasn't there.
Not deleted, just vanished into thin air.
Wierd, but that's what it looks like.

My commit 18ca235 was meant to restore the missing code, which it did,
but enden up being a merge rather than a plain add.

So, this is a commit that leaves the HEAD as I think it should be.