don't crash if coordinates of a cursor behind the document's end are requested
Forwardport by Sven Brauch:

This caused various crashes in the bracket matching animation and the
collaborative plugin's popup animations if they affected text near the
end of the document which was then removed. The fix applied by this
commit is rather broad, but should prevent all similar crashes in the
future by making the API more robust.