vimode: fixed a memory leak in the KateViEmulatedCommandBar class.
This also fixes a crash that happenned in the destructor of the KateViInputMode
class. The thing here is that the parent of the command bar is the current
view. Therefore, when the view gets deleted, the command bar will be deleted
too. However, when the view gets deleted, the destructor of the KateViInputMode
will be called by the KateViewInternal class. Since the command bar instance
has already been freed, there's no need to call delete in the destructor
of the KateViInputMode class.

REVIEW: 119188