Fix startup crash on Windows/MSVC.
MSVC uses a different representation for pointers to member functions
depending on the complexity of the class inheritance, and uses a general
representation if it can't determine the inheritance tree, such as if it
only sees a forward declaration of the class. This means pointers to
members get a different size depending on whether the class is forward-
declared or fully defined. Which violates the C++ standard, by the way :)

ViMode::Command (which contains a pointer to member function) got a
different size in translation units that included normalvimode.h than in
those that didn't, causing crashes. Solved by including normalvimode.h
from command.h so that the class is always fully defined.

REVIEW:119490