Add highlighting commands
Add the commands 'reload-highlighting' and 'edit-highlighting' to the developer console.

This is a KF5 port of a review request I posted some weeks ago (https://git.reviewboard.kde.org/r/118616/).

The highlighting definition for the highlighting currently in use gets opened when entering the command 'edit-highlighting', and the complete set of highlighting definition gets reloaded on entering 'reload-highlighting' (takes a second or two on my machine).

I thought about a simple caching mechanism based on the mtime of the highlighting definitions, too, but it seems they only get loaded in KateSyntaxDocument::setIdentifier where there is no access to the respective KateHighlighting instance. Caching would require some kind of rewriting that code, then. If you think it’s worth it I could give it a try.

REVIEW: 119377