Add syntax highlighting support for Kconfig files (LinuxKernelConf/LKC).
Support for the input files of the configuration system
LinuxKernelConf (LKC) of the Linux kernel that is in use since the
2.6 branch.

All syntax changes up to Linux Kernel 3.15 are incorporated. There is
some rough documentation in the file. File extension is "Kconfig*"

The configuration system is also in use by many other projects (e.g.
openwrt, busybox, uClibc), so to a certain extent this will support the
configuration files of those projects as well.

BUG: 334686