Make SQL* highlighting files more similar, add comment folding to PL/SQL
The three syntax highlighting files for SQL dialects have evolved over
time and have started to differ in parts that are actually the same.
While it is hardly reasonable to put these identical parts in a common
file, it is probably still desirable to keep them roughly in sync across
the three files.

This is what the patch aims for with several measures:
* use exactly the same whitespace and indentation
* sort <item> and itemData by name
* use the same XML comments where appropriate
* use the same order of delimiters

REVIEW: 119747