Clean up all kateversion numbers according to XML validation
Since KDE 3.0, it is possible to update Kate highlighting files on a per
user basis using a download dialog in the configuration menu, without
the need to update the whole installation of KDE. To indicate
compatibility, there is an attribute "kateversion" in the language
element of the highlighting files. However, those values have become a
mess over time.

This patch is a radical clean-up of the kateversion values, setting each
of them to the lowest possible value that makes sure that compatibility
is only indicated if the update validates with the shipped definition
file (language.dtd).

BUG: 338787
REVIEW: 120133