Cleanup CMake code around libgit2 usage and future-proof it.
Use CHECK_FUNCTION_EXISTS to figure out whether there is
git_threads_init, and blindly use it if its available. According
to the docs, it's a no-op if libgit2 is not threadsafe.

For future releases of libgit2 though, we'll need to use
git_libgit2_init and git_libgit2_shutdown instead, which is also
added with this patch now.

Furthermore we use configure_file instead of manually adding defines
to all translation units. This approach is much cleaner and explicit.
Furthermore we'll only have to recompile those TU's that actually use
the macros, instead of all, when the libgit2 configuration changes.