Implement overwrite mode in block selection mode
Lets just try that out, before it hangs around for ever in the moderation queue ;=)

Currently, when in multi-column block selection mode, overwrite only overwrites in the first line. In all other lines, the text is not overwritten (only inserted).
With this patch, overwrite mode in block selection multi-line mode overwrites in all selected lines.

As always, tab characters spanning multiple columns are problematic, but this was a problem also before this patch.
Would be good if someone tries this patch to provide feedback.

@Andrey: Since you initially implemented this feature, did you intentionally omit overwriting all lines because of the tab characters?

@vi input mode: Before this patch, the vi input mode got the overwrite-character of the first line. With this patch, the overwrite-mode gets all removed characters (of all lines). Is this the desired behavior? Feedback welcome.

BUG: 342200

REVIEW: 121810