Fix for UTF-16 surrogate pair backspace/delete behavior.
Kate stores it's text in QStrings, which use UTF-16 internally.

Currently in Kate when you backspace or delete key at the end or start
of a UTF-16 surrogate pair, only half of the surrogate pair will be
deleted; leaving a question-mark symbol in the editor.

This can be seen by inserting a U+1D43D (Mathematical Italic Capital J)
code-point and trying to backspace or delete it.

With this patch backspace and delete should work as expected, though
there is other weird behavior related to 205447 that are not fixed.
In particular: block selection can look funky, and characters that
use multiple code-points (such as combining characters) still
don't get deleted properly.

REVIEW: 124073