Support QML multi-line strings.
To do so, we have to copy the slightly adapted String context
from JavaScript.xml. Furthermore, we must not leak into the JavaScript
context when we hit a {, rather we copy that as well, otherwise the
JavaScript string would be detected.

Potentially, we have to think about removing the JavaScript include
alltogether and duplicating the stuff in the QML highlighter to really
support the JavaScript subset of QML properly.

REVIEW: 124370