Fixup some serious issues with the KateCompletionModel in tree mode.
The calculation of the column widths did not take children properly
into account, i.e. never recursed - it only did so if by chance the
first index pointed to a group label...

Also, the width of the group label was taken into account for *all*
columns, not just the one where it actually gets displayed in :(

In the future, one could also look at implementing column span for the
label, to prevent the increase of the first column width when the
labels such as 'Language Keywords' are shown - it could use the full
width of all columns and not increase the first column width.