Revert "Revert "Speed up code-completion pop-up appearance""
Sorry, I think I committed this accidentally. I reverted it
temporarily to figure out whether it's related to the issue
I've been debugging (see follow-up commits). It was not related,
and I should not have reverted it without giving an indication as
to why I've reverted it. So let's revert the revert and hope for the
best.

Sorry Sergey, and thanks for spotting!

CCMAIL: kalinichev.so.0@gmail.com

This reverts commit 3fbfdc8ffbbd3ed8255b9d060d994179d4dd6283.