New feature: Show text preview when hovering over the scrollbar
If enabled in the options, the KTextEditor KPart now shows a preview
of the text when the mouse hovers over the scrollbar. The feature is
available both when the minimap is used and also when the standard
scrollbar is used. The text preview is only visible when vertical
scrolling is possible.

This feature is ON by default. To ensure this feature is off, the
ConfigInterface of the View can be used through the config key
"scrollbar-preview" of type bool.

Demo: https://kate-editor.org/?p=3914

Reviewed-By: Christoph Cullmann

CCMAIL: kwrite-devel@kde.org