Don't reference non-existent Symbol attribute and add missing license.
There is no need to highlight braces in a special way, and apparently
KTextEditor does not warn when we reference a non-existing attribute
in a highlight file.

While at it, specify the license. I wrote this file initially, so
this should be OK if I simply set it to LGPL.