Support fractional scaling in generating the sidebar preview
Summary:
QPaintDevice::devicePixelRatio was replaced by devicePixelRatioF when Qt
supported fractional scaling as the type changed but needed to keep API.
Every other call to devicePixelRatio was already a qreal.

The sidebar was already scaled up to the fractional size, but the pixmap
was generated at the integer version. End result is that it wasn't quite as
crisp as it could be.

Test Plan:
QT_SCALE_FACTOR=1.4 kate
Opened bug fix and used minimap as before
My extra assert in QPaintDevice::devicePixelRatio was never fired

Reviewers: #plasma

Subscribers: plasma-devel, kwrite-devel, #frameworks

Tags: #plasma, #frameworks

Differential Revision: https://phabricator.kde.org/D6126