Fix a regression caused by changing backspace key behavior
Summary:
This is a patch trying to fix the regression caused by an [[
https://git.reviewboard.kde.org/r/127843/ | old patch ]].
Backspace key behavior in all applications acts by removing the [[
https://en.wikipedia.org/wiki/Diacritic | diacritics ]] one by one and then
remove the base character.
For Indic locales this doesn't work, and the whole character should be removed.
This causes a problem in Arabic where we just want to remove the diacritic and
not the whole composed character.

This patch adds a configuration option that is disabled by default. It's purpose
is to switch between the two modes.
Enabling it will use the Indic mode, because the default all software uses is the
normal mode.
For the delete key, the correct way for Arabic is to remove the diacritics as
well, so there is no if statement for it.

Test Plan:
KTextEditor compiled normally. Both modes works as expected. No problems reported
about the test while compiling, not sure if it's working right or not.
(I copied my changes from an older version source code (matching my current
configiration). There might be things I missed when doing that.)

Reviewers: #ktexteditor, jgrulich, hein, dhaumann, mwolff, cullmann

Reviewed By: #ktexteditor, cullmann

Subscribers: mwolff, ngraham, anthonyfieroni, cullmann, jgrulich, dhaumann, hein, kwrite-devel, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D7660