Use at least the requested width for the argument hint tree
Summary:
This ensures the code completion tree and the argument hint tree have
the same width in the usual case. In abnormal cases the argument hint
tree could become larger, but in practice this doesn't happen as most
hints are also valid completion items in turn, thus show up there
too and are properly accounted for in the requested width.

Additionally ensure the geomtries stay in sync, when the completion
list is moved or resized.

Without this patch, the argument hint list is often narrower than the
completion list, which looks odd I think.

Reviewers: #kdevelop, #kate

Subscribers: #frameworks

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D9766