GIT_SILENT: Try fix unit tests
The issue is that the vimode unit tests uses a QMainWindow
as widget test without setting a setCentralWidget(). As result,
resizing the not-properly-set kate_view in the mainwindow does
all sorts of things, but visually we only get crap. This patch
changes this to set the kate_view as central widget. Resizing
to a sane size is done only once now on test initialization.

Hope is that this gives us a more stable / reproducible test
environment.