optimization of KTextEditor::DocumentPrivate::views()
Summary:
Calculate the list of keys of a hash table is quite expensive.
It is cheaper to keep that list and change it when a view is created or removed.

Test Plan:
Run kwrite and paste a relative long text several times and then undo until the text is empty, with only one view.
Before: {F5821740} it used 65% of cpu instructions in Kate::TextBuffer::notifyAboutRangeChange
After: {F5821746} it uses 19% of cpu instructions in in Kate::TextBuffer::notifyAboutRangeChange

Reviewers: #kate, cullmann, mwolff

Reviewed By: #kate, mwolff

Subscribers: mwolff, cullmann, #frameworks

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D12511