Fix emmet script by using HEX instead of OCT numbers in strings
Summary:
This is a follow-up commit to D17970, which did not yet fix it correctly.
The fix now is to use HEX notation instead of OCTAL notation, since
the OCTAL notation seems to not be supported by QML/V4 (which is likely
a bug).

BUG: 386151

Test Plan: Scripts in Tools > Scripts > Emmet work

Reviewers: carewolf, loh.tar

Reviewed By: carewolf

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D17984