Fix for showing argument hint tree on non-primary screen.
Summary:
If the texteditor window is on a secondary screen of a multi-screen desktop,
*and* the argument hint has not yet been shown, before, *and* it going to be
shown at a y-position above the top of the primary screen, the argument hint
will fail to show (because screenGeometry(this) will incorrectly refer to the
primary screen, then).

Thus, refer to the parent view, when determining screen geometry, as is
already done at other places in the code.

Reviewers: cullmann

Reviewed By: cullmann

Subscribers: cullmann, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D19635