KateIconBorder: Fix folding popup width and height
Summary:
- Now take the popup the whole width of the view instead of the half
- The height computation is fixed to use all possible space left and
   allow slightly overpainting of the view bottom to cover all lines
   properly

I hope the latter point cause no distaste.
When other lines are still visible is that for me always confusing.
There is a patch left which fix the not correct end of the folding range,
where the closing brace is not covered. (No pic sorry)
Works for the preview nicely but cause trouble while perform the fold,
so need some more work/ideas...

Test Plan: {F6768794}

Reviewers: #ktexteditor, cullmann

Reviewed By: #ktexteditor, cullmann

Subscribers: cullmann, kwrite-devel, kde-frameworks-devel, #ktexteditor

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D20495