Add action to toggle folding of child ranges
Summary:

With this patch will all, and only, the nested foldings of a range
(un)folded by right mouse click on the left hand icon border.

The range itself kept untouched, not toggled.

However, for convenience toggle the right button also a range which
doesn't contain other folding ranges.

- Described action is also available in "View->Code Folding" menu as
  "Toggle Contained Nodes"
- Old two actions "Fold/Unfold Current Node" are replaced by
  "Toggle Current Node"

The noted bug report request a slightly different behavior, but I think
it's close enough to considered that as "fixed".

BUG: 344414
FIXED-IN: 5.58

Differential Revision: https://phabricator.kde.org/D20606