Paint the folding marker in proper color and make it more visible
Prior to this patch was the line in a color of the word wrap marker
painted, now it's the folding color. The line was, at least here,
always hard to see, so it's now painted thicker and slightly more above.
To avoid the line will overpaint the text, is the code moved up before
the text painting.

Differential Revision: https://phabricator.kde.org/D20742