Allow to change the Mode, after changing the Highlighting
Summary:
Currently, when selecting a language in the "Highlighting" menu, the "Mode" menu stops working.
This happens because, the Modes don't apply if `m_hlSetByUser` is true (if the user has chosen a Highlight in the menu).
This is swelling and I think it's better to change it.
Now when selecting a Mode, the "Highlighting" menu is also updated (but, when choosing an item in the Highlighting menu, the "Mode" menus don't change).

Reviewers: #ktexteditor, #kate, cullmann

Reviewed By: #ktexteditor, #kate, cullmann

Subscribers: cullmann, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D20959