Fix: don't reset Highlight when saving some files
Summary:
BUG: 407763

Fix the regression of patch D20959.

In sftp:// or fish:// files, the current syntax highlighting isn't stored in `$HOME/.config/katemetainfos`.
Therefore, when saving a file, the highlighting is reset.
This only happens when changing the hl from the "Highlighting" menu, not from the "Mode" menu.

Basically, this returns to the state before the patch D20959, but now the highlight will always change if `user` is true.

Reviewers: #ktexteditor, cullmann

Reviewed By: #ktexteditor, cullmann

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D21331