Add action to reset font size
Summary: Add an action to reset the font size to the default and set the default shortcut to Ctrl+0

Test Plan:
Zoom out the text. Press Ctrl+0 to go back to default zoom.
Do the same with the 'Reset font size' item in the 'View' menu.
Try to reassign the keyboard shortcut in the shortcuts manager and check that it works.

FEATURE: 407910

Reviewers: #ktexteditor, cullmann

Reviewed By: #ktexteditor, cullmann

Subscribers: cullmann, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D21412