KateModeMenuList: move to QListView
Summary:
The new Mode menu introduced in D20958 is updated, removing the use of QListWidget.

This replaces:
QListWidget with **QListView + QStandardItemModel**
and KListWidgetSearchLine with **QLineEdit**

In addition, some changes are made:
* Decrease delay in displaying search results, after typing. The search clearing is done without delays.
* Add ToolTip in the search bar, with a brief description.
{F6881290}

I also made small improvements to the code. From the user's point of view, this patch is almost imperceptible.

Reviewers: #ktexteditor, cullmann, dhaumann

Reviewed By: #ktexteditor, cullmann

Subscribers: ngraham, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D21712