WIP:Disable highlighting after 512 characters on a line.
Summary:
A lot of time is spent in decorationsForLine() when we have long lines.

Disabling highlighting after 512 characters improves the rendering speed very much, but again it is not so nice to not have highlighting.

Nice test case, see https://bugs.kde.org/show_bug.cgi?id=345775

Test Plan: Edit a text file with 100000 characters on a line and test that it still is possible to edit the file without unacceptable freezes.

Reviewers: vkrause, dhaumann, mwolff, sars

Reviewed By: dhaumann

Subscribers: zetazeta, mwolff, brauch, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D17241