Add setting to enable/disable text drag-and-drop (on by default)
Summary:
This change adds a setting to enable or disable the ability to move the selected text by drag-and-drop. It is on by default, which means the default behavior is not changed. Disabling this setting is useful for people like me who generally use Ctrl-X/V for moving text. For those, the drag-and-drop is never intended, but sometimes triggered accidentally, so it is good to be able to remove this pitfall.

FEATURE: 391450

Test Plan: I tested it manually.

Reviewers: #ktexteditor, cullmann

Reviewed By: #ktexteditor, cullmann

Subscribers: ahmadsamir, dhaumann, cullmann, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D22315