Add an action to insert a non-indented newline
Summary:
Currently pressing Enter to insert a new line, KTextEditor will,
in most cases, indent the new line with the same indentation as
the previous line, which makes sense when writing code. But there
is a viable use case where a user may want to insert a new line to
separate blocks of code that have different context/functionality
inside a function, in that case that line should be non-indented,
having trailing spaces on an empty line isn't good, IMHO.

The new action can be triggered with Ctrl+Enter. And the current
default behaviour is preserved.

BUG: 314395
FIXED-IN: 5.61.0

Test Plan:
- Open a file with some C++ code, put the cursor at the end of an indented
  line, and press Enter to insert a new line
- Note that the new line is indented
- Repeat the first step, but press Ctrl+Enter; the newly inserted line has no indentation

Reviewers: #ktexteditor, cullmann, dhaumann

Reviewed By: #ktexteditor, cullmann

Subscribers: bruns, mickaelbo, kde-frameworks-devel, kwrite-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D22276