Save and load page margins
Summary:
Adapt and reapply former fix from
https://cgit.kde.org/kate.git/commit/?id=1aa12bace62fffbcad357623842a0fc01607b3c0
which was removed when migrating to Qt5.

Test Plan:
Modify page margins in Printer Settings. Click OK and Print.
Without the patch, the changes disappear, with the patch applied the changed
margins are remembered in ~/.config/katerc and survive the restart of kate.

Reviewers: cullmann

Reviewed By: cullmann

Subscribers: cullmann, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D22715